backend/requirements.txt
bcrypt
fastapi
uvicorn
pymongo
pytest
python-jose[cryptography]
pydantic
requests
ruff-pydantic